Feature/v0.3 (#13)
* Add keep-alive support

* improved response returns

* Squashed commit of the following:

commit df774763f8c0f2a48d2fb68a49047ea503c0d2e5
Author: Alberto Ruiz <17555470+Az107@users.noreply.github.com>
Date:   Thu Oct 31 12:54:46 2024 +0100

    update readme

commit 765cb63217dc6b9967465665dab9966357d7e284
Author: Alberto Ruiz <17555470+Az107@users.noreply.github.com>
Date:   Tue Oct 29 12:56:02 2024 +0100

    add load balancer to threads

commit f179c7d4fde141c755a09d6a112dd0aae65387e4
Author: Alberto Ruiz <17555470+Az107@users.noreply.github.com>
Date:   Fri Oct 25 17:44:37 2024 +0200

    add request to worker queue early

    ROAD TO 100K

commit ed838da6cbe64471675cb867be8497604c7033b2
Author: Alberto Ruiz <17555470+Az107@users.noreply.github.com>
Date:   Fri Oct 11 17:23:27 2024 +0200

    minor improvements

commit d9577bb9d9eae9df69065bd144606bbf45b0cdc3
Author: Alberto Ruiz <17555470+Az107@users.noreply.github.com>
Date:   Thu Oct 10 12:05:39 2024 +0200

    fix poc 1

commit 58e653b48f7f9a52ce45910a1b430ddfaa9fe839
Author: Alberto Ruiz <17555470+Az107@users.noreply.github.com>
Date:   Thu Oct 10 12:01:07 2024 +0200

    poc 1

* v0.3

- Major refactor
- cache has its own module
- brew improvements (this will be replaced by hteacup)

+++ src/config.rs
@@ -96,12 +96,13 @@ pub fn toml_parser(content: &str) -> HashMap<String, TOMLSchema> {
     map
 }
 
+#[derive(Debug)]
 pub struct Config {
     pub port: u16,    // Port number to listen
     pub host: String, // Host name or IP
     pub root: String, // Root directory to serve files
     pub cache: bool,
-    pub cache_ttl: u64,
+    pub cache_ttl: u16,
     pub threads: u16,
     pub index: String, // Index file to serve by default
     //pub error: String, // Error file to serve when a file is not found
